apps/workers/requirements.txt
# GlassBox Python Workers Dependencies

# AWS SDK
aioboto3>=12.0.0
boto3>=1.34.0

# Database
asyncpg>=0.29.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging
structlog>=24.0.0

# LLM
litellm>=1.30.0

# Text extraction
pypdf>=4.0.0
python-docx>=1.0.0
aiofiles>=23.0.0

# Utils
python-dotenv>=1.0.0
